openmp RC 1
- Separated definitions and implementations
- Separated parts of code to different files
- Added matrix text files for testing
- Fixed an an issue in utils.cpp SwapRows (used rows instead of cols)
- Changed second loop starting point to diagonal
- Added algorithm choosing to main
- Fixed other minor issues in main
- Added OpenMP implementation
