+++ README.md
@@ -1,0 +1,17 @@
+# Matrix inversion program (Gaussian elimination)
+
+To use compile with `g++ *.cpp -o main -fopenmp -I .`
+
+## Inputs:
+- Accepts input from "Input.txt" file in current directory (first number in "Input.txt" is matrix size)
+- Generates a random matrix with chosen size
+
+To make file inputs easier there are a few matrices prepared.
+
+## Outputs:
+- Outputs to "Output.txt" file in current directory
+- Console
+
+## Algorithm types:
+- Sequential
+- Parallel using OpenMP
