--- README.txt
@@ -1,6 +1,0 @@
-Matrix inversion program (Gaussian elimination) by Jakub Maliszewski MZ02IP1
-
-Accepts input from "Input.txt" file in current directory or generates a random matrix with chosen size.
-First number in "Input.txt" is matrix size.
-
-Outputs either to "Output.txt" file in current directory or console.
