+++ matrix_10x10.txt
@@ -1,0 +1,11 @@
+10
+2 7 2 9 4 5 0 0 5 5
+7 4 2 1 4 8 0 2 4 5
+6 4 6 8 9 4 3 0 3 2
+6 9 0 6 2 1 6 1 4 6
+8 4 9 2 5 5 2 6 9 2
+9 1 2 4 6 1 5 8 2 3
+1 6 7 5 1 0 4 8 0 8
+5 5 9 0 4 6 2 5 8 7
+9 7 2 6 7 4 4 0 9 6
+6 9 6 4 9 4 6 5 1 0
