+++ matrix_3x3.txt
@@ -1,0 +1,4 @@
+3
+2	-1	2
+-1	3	0
+1	2	3
