+++ openmp.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <utils.h>
+
+// Perform Gaussian elimination to find matrix solution with OpenMP
+void OpenMPGaussianElimination(Matrix2D<double> &matrix);
